+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Documentation, examples and support for Jobs API 2.1
 - Support for Jinja2-based templates inside deployment configuration
+- Added new `--job` argument to deploy command for a single-job deploy and convenience
 
 ### Fixed